Fix Cast endoding and decoding

+++ zinc-compiler/src/semantic/analyzer/expression.rs
@@ -351,7 +351,7 @@ impl Analyzer {
                             .borrow_mut()
                             .push_instruction(Instruction::Cast(zinc_bytecode::Cast::new(
                                 is_signed,
-                                bitlength as u8,
+                                bitlength,
                             )));
                     }
